FileUploadProcessor refactoring
