+++ DeliveryProject.DataAccess/Processors/FileUploadProcessor.cs
@@ -1,22 +1,21 @@
-﻿using DeliveryProject.Bussiness.Interfaces.Services;
-using DeliveryProject.Core.Enums;
-using DeliveryProject.Core.Models;
+﻿using DeliveryProject.DataAccess.Entities;
 using DeliveryProject.Core.Extensions;
 using DeliveryProject.DataAccess.Interfaces;
+using DeliveryProject.DataAccess.Enums;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using DeliveryProject.Core.Constants.InfoMessages;
 using DeliveryProject.Core.Constants.ErrorMessages;
 
-namespace DeliveryProject.Bussiness.Services
+namespace DeliveryProject.DataAccess.Processors
 {
-    public class FileUploadService : IFileUploadService
+    public class FileUploadProcessor : IFileUploadProcessor
     {
         private readonly IBatchUploadRepository _batchUploadRepository;
         private readonly string _uploadFolder = "uploads";
-        private readonly ILogger<FileUploadService> _logger;
+        private readonly ILogger<FileUploadProcessor> _logger;
 
-        public FileUploadService(IBatchUploadRepository batchUploadRepository, ILogger<FileUploadService> logger)
+        public FileUploadProcessor(IBatchUploadRepository batchUploadRepository, ILogger<FileUploadProcessor> logger)
         {
             _batchUploadRepository = batchUploadRepository;
             _logger = logger;
@@ -39,15 +38,19 @@ namespace DeliveryProject.Bussiness.Services
 
             _logger.LogInformation(BatchUploadInfoMessages.FileIsSaved, file.FileName, filePath);
 
-            var batchUpload = new BatchUpload
+            var batchUpload = CreateBatchUpload(file, filePath, uploadType);
+            await _batchUploadRepository.AddAsync(batchUpload);
+            return batchUpload;
+        }
+
+        private BatchUpload CreateBatchUpload(IFormFile file, string filePath, UploadType uploadType)
+        {
+            return new BatchUpload
             {
                 FileName = file.FileName,
                 FilePath = filePath,
                 UploadType = uploadType
             };
-
-            await _batchUploadRepository.AddAsync(batchUpload);
-            return batchUpload;
         }
     }
 }
